+++ egui/src/widgets/mod.rs
@@ -53,7 +53,7 @@ pub fn reset_button<T: Default + PartialEq>(ui: &mut Ui, value: &mut T) {
 pub fn stroke_ui(ui: &mut crate::Ui, stroke: &mut epaint::Stroke, text: &str) {
     let epaint::Stroke { width, color } = stroke;
     ui.horizontal(|ui| {
-        ui.add(DragValue::f32(width).speed(0.1).range(0.0..=5.0))
+        ui.add(DragValue::f32(width).speed(0.1).clamp_range(0.0..=5.0))
             .on_hover_text("Width");
         ui.color_edit_button_srgba(color);
         ui.label(text);
@@ -70,8 +70,12 @@ pub(crate) fn shadow_ui(ui: &mut Ui, shadow: &mut epaint::Shadow, text: &str) {
     let epaint::Shadow { extrusion, color } = shadow;
     ui.horizontal(|ui| {
         ui.label(text);
-        ui.add(DragValue::f32(extrusion).speed(1.0).range(0.0..=100.0))
-            .on_hover_text("Extrusion");
+        ui.add(
+            DragValue::f32(extrusion)
+                .speed(1.0)
+                .clamp_range(0.0..=100.0),
+        )
+        .on_hover_text("Extrusion");
         ui.color_edit_button_srgba(color);
     });
 }